fix: 커스텀 구분자 오류 처리
커스텀 구분자로 이스케이프 문자가 들어오는 경우에도 정상적으로 작동되도록 수정
계산기에 매우 큰 값이 들어올 경우를 대비하여 인자 및 결과값을 int에서 long으로 수정

+++ src/main/java/calculator/Application.java
@@ -8,7 +8,7 @@ public class Application {
             System.out.println("덧셈할 문자열을 입력해 주세요.");
             String userInput = Console.readLine();
             StringCalculator calculator = new StringCalculator();
-            int result = calculator.calculate(userInput);
+            long result = calculator.calculate(userInput);
             System.out.println("결과 : " + result);
         } catch (IllegalArgumentException e) {
             System.out.println("잘못된 문자열을 입력하였습니다!!!");
